Move berry puch back (not a held item)
