+++ src/modifier/modifier.ts
@@ -1281,6 +1281,42 @@ export class CriticalCatchChanceBoosterModifier extends PersistentModifier {
   }
 }
 
+export class PreserveBerryModifier extends PersistentModifier {
+  match(modifier: Modifier) {
+    return modifier instanceof PreserveBerryModifier;
+  }
+
+  clone() {
+    return new PreserveBerryModifier(this.type, this.stackCount);
+  }
+
+  /**
+   * Checks if all prequired conditions are met to apply {@linkcode PreserveBerryModifier}
+   * @param pokemon {@linkcode Pokemon} that holds the berry
+   * @param doPreserve {@linkcode BooleanHolder} that is `true` if the berry should be preserved
+   * @returns `true` if {@linkcode PreserveBerryModifier} should be applied
+   */
+  override shouldApply(pokemon?: Pokemon, doPreserve?: BooleanHolder): boolean {
+    return !!pokemon && !!doPreserve;
+  }
+
+  /**
+   * Applies {@linkcode PreserveBerryModifier}
+   * @param pokemon The {@linkcode Pokemon} that holds the berry
+   * @param doPreserve {@linkcode BooleanHolder} that is `true` if the berry should be preserved
+   * @returns always `true`
+   */
+  override apply(pokemon: Pokemon, doPreserve: BooleanHolder): boolean {
+    doPreserve.value ||= pokemon.randBattleSeedInt(10) < this.getStackCount() * 3;
+
+    return true;
+  }
+
+  getMaxStackCount(): number {
+    return 3;
+  }
+}
+
 export class LockModifierTiersModifier extends PersistentModifier {
   match(modifier: Modifier): boolean {
     return modifier instanceof LockModifierTiersModifier;
