+++ src/phases/turn-start-phase.ts
@@ -5,7 +5,7 @@ import { Abilities } from "#app/enums/abilities";
 import { Stat } from "#app/enums/stat";
 import type Pokemon from "#app/field/pokemon";
 import { PokemonMove } from "#app/field/pokemon";
-import { BypassSpeedChanceModifier } from "#app/modifier/modifier";
+import { BypassSpeedChanceModifier } from "#app/modifier/held-item-modifier";
 import { Command } from "#app/ui/command-ui-handler";
 import { randSeedShuffle, BooleanHolder } from "#app/utils/common";
 import { AttemptCapturePhase } from "./attempt-capture-phase";
