+++ test/abilities/harvest.test.ts
@@ -1,7 +1,8 @@
 import { BattlerIndex } from "#app/battle";
 import { PostTurnRestoreBerryAbAttr } from "#app/data/abilities/ability";
 import type Pokemon from "#app/field/pokemon";
-import { BerryModifier, PreserveBerryModifier } from "#app/modifier/held-item-modifier";
+import { BerryModifier } from "#app/modifier/held-item-modifier";
+import { PreserveBerryModifier } from "#app/modifier/modifier";
 import type { ModifierOverride } from "#app/modifier/modifier-type";
 import type { BooleanHolder } from "#app/utils/common";
 import { Abilities } from "#enums/abilities";
